+++ DocStoreAPI/Controllers/BuisnessAreaController.cs
@@ -15,6 +15,8 @@ namespace DocStoreAPI.Controllers
     {
         private readonly BuisnessAreaRepository _buisnessAreaRepository;
         private readonly SecurityRepository _securityRepository;
+        private const string _objectType = "BuisnessArea";
+
         public BuisnessAreaController(BuisnessAreaRepository repository, SecurityRepository securityRepository)
         {
             _buisnessAreaRepository = repository;
@@ -24,6 +26,9 @@ namespace DocStoreAPI.Controllers
         [HttpGet]
         public IActionResult Get()
         {
+            if (!_securityRepository.UserIsAdmin(HttpContext))
+                return _securityRepository.GateUnathorised(HttpContext, AccessLogAction.BAList, _objectType, "NA");
+
             var buisnessAreas = _buisnessAreaRepository.List();
 
             return Ok(buisnessAreas);
@@ -33,15 +38,24 @@ namespace DocStoreAPI.Controllers
         [HttpGet("{id}", Name = "Get")]
         public IActionResult Get(int id)
         {
-            var ba = _buisnessAreaRepository.GetById(id);
+            var entity = _buisnessAreaRepository.GetById(id);
 
-            return Ok(ba);
+            if (String.IsNullOrWhiteSpace(entity.Name))
+                return _securityRepository.GateNotFound(HttpContext, AccessLogAction.BAReturn, _objectType, id.ToString());
+
+            if (!_securityRepository.UserIsAdmin(HttpContext))
+                return _securityRepository.GateUnathorised(HttpContext, AccessLogAction.BAReturn, _objectType, id.ToString());
+
+            return Ok(entity);
         }
 
         // POST: api/BuisnessArea
         [HttpPost]
         public IActionResult Post([FromBody] BuisnessAreaEntity value)
         {
+            if (!_securityRepository.UserIsAdmin(HttpContext))
+                return _securityRepository.GateUnathorised(HttpContext, AccessLogAction.BACreate, _objectType, "NA");
+
             _buisnessAreaRepository.Add(value);
             _buisnessAreaRepository.SaveChanges();
 
@@ -52,6 +66,14 @@ namespace DocStoreAPI.Controllers
         [HttpPut("{id}")]
         public IActionResult Put(int id, [FromBody] BuisnessAreaEntity value)
         {
+            var entity = _buisnessAreaRepository.GetById(id);
+
+            if (String.IsNullOrWhiteSpace(entity.Name))
+                return _securityRepository.GateNotFound(HttpContext, AccessLogAction.BAUpdate, _objectType, id.ToString());
+
+            if (!_securityRepository.UserIsAdmin(HttpContext))
+                return _securityRepository.GateUnathorised(HttpContext, AccessLogAction.BAUpdate, _objectType, id.ToString());
+
             _buisnessAreaRepository.Edit(value);
             _buisnessAreaRepository.SaveChanges();
 
@@ -62,6 +84,14 @@ namespace DocStoreAPI.Controllers
         [HttpDelete("{id}")]
         public IActionResult Delete(int id)
         {
+            var entity = _buisnessAreaRepository.GetById(id);
+
+            if (String.IsNullOrWhiteSpace(entity.Name))
+                return _securityRepository.GateNotFound(HttpContext, AccessLogAction.BADelete, _objectType, id.ToString());
+
+            if (!_securityRepository.UserIsAdmin(HttpContext))
+                return _securityRepository.GateUnathorised(HttpContext, AccessLogAction.BADelete, _objectType, id.ToString());
+
             _buisnessAreaRepository.DeleteById(id);
             _buisnessAreaRepository.SaveChanges();
